requirements.txt
boto3
matplotlib
scikit-learn
numpy
tiktoken
pandas
faiss-cpu
sentence-transformers
html2text
beautifulsoup4
mcp